add xml docs
