+++ JsonSchema.Generation.Tests/JsonSchema.Generation.Tests.csproj
@@ -10,7 +10,7 @@
 		<RootNamespace>Json.Schema.Generation.Tests</RootNamespace>
 		<SignAssembly>true</SignAssembly>
 		<AssemblyOriginatorKeyFile>../json-everything.snk</AssemblyOriginatorKeyFile>
-		<NoWarn>IDE1006</NoWarn>
+		<NoWarn>IDE1006,CS8618</NoWarn>
 	</PropertyGroup>
 
 	<ItemGroup>
