+++ JsonSchema.Generation/Attributes/RequiredAttribute.cs
@@ -9,5 +9,7 @@ namespace Json.Schema.Generation;
 [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field, AllowMultiple = true)]
 public class RequiredAttribute : ConditionalAttribute
 {
+#pragma warning disable CS8618
 	internal string PropertyName { get; set; }
+#pragma warning restore CS8618
 }
