+++ JsonSchema.Generation/Intents/IfIntent.cs
@@ -2,8 +2,14 @@
 
 namespace Json.Schema.Generation.Intents;
 
+/// <summary>
+/// Provides intent to create an `if` keyword.
+/// </summary>
 public class IfIntent : ISchemaKeywordIntent
 {
+	/// <summary>
+	/// A set of intents used to define the subschema.
+	/// </summary>
 	public IEnumerable<ISchemaKeywordIntent> Subschema { get; }
 
 	/// <summary>
